doc/requirements.txt
Flask==3.0.3
GitPython==3.1.43
javalang==0.13.0
requests==2.32.3
APScheduler==3.10.4
